refactor(db): extract committed checkpoint index
